Revert to default upstream behavior for determining maxPoolSize
We'll need to set INVOKER_NUMCORE and INVOKER_CORESHARE environment
variables appropriately during deployment if we want to run more than
the default of 8 containers per Invoker.

The correct numbers will depend on the size of the Kubernetes cluster
but MAX_POOL_SIZE = INVOKER_NUMCORE * INVOKER_CORESHARE so we'll want
one or both of these numbers to be higher.