Reduce shelling out to `kubectl` in KubernetesClient
Combine creating a pod and labeling it into a single command to go
from 2 to 1 invocations of `kubectl` for each new pod created.

Use the Fabric8 Kubernetes client instead of `kubectl` to wait for
pods to become ready and fetch their IP address. This also increases
the default timeout substantially to handle cases where Kubernetes is
overloaded and takes longer to create pods.