Add unit tests for Kubernetes{Container,Client}
These mostly match the relevant analog tests for Docker.

A number of refactorings came out of the tests, namely
the run methods: moving the setting of the kubectl flags
from the Client to the Container.

We're also no longer hard-coding the runtime memory limit;
it defaults to its previous value.