Use Kube labels to indicate status of function containers
When using the KubernetesContainer implementation, function pods will
now be labeled to indicate their status of prewarmed, active, or
paused. We're also labeling the function pods with the name of the
container image they use to make it easy for an OpenWhisk
administrator to filter function pods by node/java/python/etc in
`kubectl`, `oc`, or the OpenShift web console.